shift part simple_nonlinear material model from test to a separate benchmark folder called nonlinear_channel_flow.
